change render photo
